Prevent circular dependency and ensure correct factories called
